+++ packages/maplibre_ios/CHANGELOG.md
@@ -1,3 +1,14 @@
+## Unreleased
+
+### Bug Fixes
+
+- iOS: fix `animateCamera()` crash when called before the native view has a layout size
+- iOS: fix `MLNMapView` not torn down on dispose
+
+## 0.3.6
+
+[Check the maplibre changelog](https://pub.dev/packages/maplibre/changelog#036)
+
 ## 0.3.5
 
 [Check the maplibre changelog](https://pub.dev/packages/maplibre/changelog#035)